Updated psrfits.py
This routine is pulled from presto (where I'm having the same issue). I have 4bit AO search mode data, and it has been stalling on this part. That's because the shp list has a length of 1 for this data. I think this is the proper modification but check me on it :D